# swagger/swagger.yaml
openapi: 3.0.0
info:
  version: '0.0.1'
  title: "HUFSPace API Documents"
  description: "한국외국어대학교 웹 페이지 프로젝트 API 문서"
servers:
  - description: "테스트 서버"
  - url: "http://52.78.2.40:8080/"
paths:
  /user:
    get:
      summary: 회원 정보 조회
      consumes: application/json
      produces: application/json
      parameters:
      responses:
        200:
          description: 회원 정보 정상 조회
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    example: {"email":"0417taehyun@gmail.com","webMail":"dlxogus0417@hufs.ac.kr","nickname":"도비","mainMajor":"스칸디나비아어과","doubleMajor":"전자공학과","myPost":[{"id":1,"title":"첫 게시글"},{"id":2,"title":"두번째 게시글"},{"id":3,"title":"졸업생 게시판 첫 번째 게시글"}],"myReplies":[{"id":1,"content":"첫 번째 게시글에 대한 첫 댓글","Post":{"id":1,"title":"첫 게시글"}},{"id":2,"content":"두 번째 게시글에 대한 두 번째 댓글","Post":{"id":2,"title":"두번째 게시글"}},{"id":3,"content":"졸업생 게시판 첫 번째 게시글에 대한 첫 댓글","Post":{"id":3,"title":"졸업생 게시판 첫 번째 게시글"}}]}
                  message:
                    type: string
                    example: ""
        401:
          description: 로그인하지 않은 사용자
          content:
            application/json:
              schema:
                $ref: "#/definitions/Unauthorized-response"           

    put:
      summary: 회원 정보 수정
      consumes: application/json
      produces: application/json
      parameters:
        - in: body
          name: body
          description: 회원 정보 수정
          required: true
          schema:
            type: object
            required:
              - nickname
              - mainMajorId
              - doubleMajorId
            properties:
              nickname:
                type: string
                example: 이태현별명
              mainMajorId:
                type: integer
                example: 1
              doubleMajorId:
                type: integer
                example: 1              
      responses:
        200:
          description: 회원 정보 정상 수정
          content:
            application/json:
              schema:
                $ref: "#/definitions/Success-response"
        400:
          description: 닉네임을 변경한지 30일이 지나지 않았을 경우
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    example: ""
                  message:
                    type: string
                    example: INVALID_NICKNAME_TIME                   
        401:
          description: 로그인하지 않은 사용자
          content:
            application/json:
              schema:
                $ref: "#/definitions/Unauthorized-response"
        409:
          description: 이미 존재하는 닉네임
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/definitions/Conflict-nickname-response"
                  - $ref: "#/definitions/Conflict-main-major-response"
                  - $ref: "#/definitions/Conflict-double-major-response"
              examples:
                nickname-conflict:
                  summary: 이미 존재하는 닉네임
                  value:
                    data: ""
                    message: "CONFLICT_NICKNAME"
                main-major-conflict:
                  summary: 주전공을 이미 한 번 수정한 경우
                  value:
                    data: ""
                    message: "CONFLICT_MAIN_MAJOR"
                double-major-conflict:
                  summary: 이중전공을 이미 한 번 수정한 경우
                  value:
                    data: ""
                    message: "CONFLICT_DOBULE_MAJOR"
        403:
          description: 접근 권한 오류
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/definitions/Forbidden-suspension-response"
                  - $ref: "#/definitions/Forbidden-before-response"
              examples:
                suspended-user:
                  summary: 정지된 사용자
                  value:
                    data: ""
                    message: FORBIDDEN_SUSPENSION
                email-unauthorized-user:
                  summary: 이메일 인증이 되지 않은 사용자
                  value:
                    data: ""
                    message: FORBIDDEN_BEFORE            

    delete:
      summary: 회원 탈퇴
      consumes: application/json
      produces: application/json
      responses:
        200:
          description: 회원 정상 탈퇴
          content:
            application/json:
              schema:
                $ref: "#/definitions/Success-response"
        401:
          description: 로그인하지 않은 사용자
          content:
            application/json:
              schema:
                $ref: "#/definitions/Unauthorized-response"
...
  /user/sign-up:
    post:
      summary: 회원 가입
      consumes: application/json
      produces: application/json
      parameters:
        - in: body
          name: body
          description: 회원 가입에 필요한 데이터 (webMail의 경우 @hufs.ac.kr 앞 부분까지만 입력)
          required: true
          schema:
                type: object
                required:
                  - nickname
                  - webMail
                  - mainMajorId
                  - doubleMajorId
                  - isAggred
                properties:
                  nickname:
                    type: string
                    example: 도비
                  webMail:
                    type: string
                    example: dlxogus0417
                  mainMajorId:
                    type: integer
                    example: 1
                  doubleMajorId:
                    type: integer
                    example: 2
                  isAggred:
                    type: boolean
                    example: true
      responses:
        200:
          description: 회원 가입 정상 완료
          content:
            application/json:
              schema:
                $ref: "#/definitions/Success-response"
        401:
          description: 개인 정보 수집 동의를 하지 않았을 경우
          content:
            application/json:
              schema:
                $ref: "#/definitions/Unauthorized-response"        
        409:
          description: 이미 존재하는 닉네임
          content:
            application/json:
              schema:
                $ref: "#/definitions/Conflict-response"
...
definitions:
  Success-response:
    type: object
    properties:
      data:
        type: object
        example: ""
      message:
        type: string
        example: ""
  Conflict-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: CONFLICT
  Conflict-nickname-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: CONFLICT_NICKNAME
  Conflict-main-major-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: CONFLICT_MAIN_MAJOR
  Conflict-double-major-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: CONFLICT_DOBULE_MAJOR        
...
  Unauthorized-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: UNAUTHORIZED
...
  Forbidden-before-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: FORBIDDEN_BEFORE
  Forbidden-suspension-response:
    type: object
    properties:
      data:
        example: ""
      message:
        type: string
        example: FORBIDDEN_SUSPENSION
...